+++ src/Core/IDataLoader.cs
@@ -16,7 +16,16 @@ namespace GreenDonut
     /// with different access permissions and consider creating a new instance
     /// per web request. -- facebook
     /// </summary>
-    public interface IDataLoader { }
+    public interface IDataLoader
+    {
+        /// <summary>
+         /// Dispatches one or more batch requests.
+         /// In case of auto dispatching we just trigger an implicit dispatch
+         /// which could mean to interrupt a wait delay. Whereas in a manual
+         /// dispatch scenario it could mean to dispatch explicitly.
+         /// </summary>
+        Task DispatchAsync();
+    }
 
     #endregion